twin_engine_backend/requirements.txt
# --- Core Backend ---
django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=3.1,<4.0
python-dotenv>=1.2,<2.0

# --- AI & Computer Vision ---
opencv-python>=4.13,<5.0
huggingface_hub>=1.4,<2.0
openai>=2.20,<3.0
azure-ai-contentsafety>=1.0,<2.0

# --- Real-time & WebSockets ---
channels[daphne]>=4.3,<5.0
channels-redis>=4.3,<5.0

# --- Media & Streaming ---
cloudinary>=1.44,<2.0
yt-dlp>=2026.2

# --- Reporting & Analytics ---
reportlab>=4.4,<5.0
requests>=2.32,<3.0
pillow>=12.1,<13.0

# --- Machine Learning & Predictions ---
scikit-learn>=1.6,<2.0
pandas>=2.2,<3.0
numpy>=2.2,<3.0
joblib>=1.4,<2.0

# --- API Documentation ---
drf-spectacular>=0.28.0,<0.29.0

# --- Background Tasks & Scheduling ---
celery>=5.4.0,<6.0
redis>=5.0.0,<6.0
django-celery-beat>=2.6.0,<3.0
django-celery-results>=2.5.1,<3.0

# --- Security & Rate Limiting ---
django-axes>=8.0,<9.0

# --- Deployment & Utilities ---
whitenoise>=6.11,<7.0
gunicorn>=25.0,<26.0
